Added MAP and JSON files
Added maps for all 8 levels. Added JSON files for all levels except for 6,
7, and 8.
The JSON files may not be in the correct format. This needs to be edited.
Also JSON files need to be edited to allow for the multiple monster waves
along with the multiple fruit waves in levels 5 and 6.
The Levels.py file was used for translating the hardcoded location code in
the master iSCREAM.py into the grid for MAP files.